+++ src/pkg/math/pow.go
@@ -98,7 +98,7 @@ func Pow(x, y float64) float64 {
 	}
 
 	// ans = a1 * 2**ae (= 1 for now).
-	a1 := float64(1)
+	a1 := 1.0
 	ae := 0
 
 	// ans *= x**yf
